Number of required sorting passes can be lowered
LSB radix sort is actually what we want (and what we have) since we sort the highest bits last, meaning they have the highest priority.
We use distance to camera instead because it seems more stable, that might not be case on mobile, where the camera position also has small frequent changes.