Now able to change materials
Commandbuffers do not work properly with different depth buffers neither from camera nor textures.
DrawTexture makes particles look different, couldn't find a way to change it. Had to use tex2Dlod sampling to avoid the particles getting darker when further away? Maybe I just need to set a different material on DrawTexture. The texture itself is the same as when not rendering to a texture.
Also experimented with zwrite enabled sphere rendering, where we simply discard the parts of the point texture that have an alpha value below a certain threshold.