Added more efficient billboard texture
Since we're are only using alpha values greater than 0.7, we can take much better advantage of the billboard texture.
A version of the billboard texture has been made, where it's half the size, by removing the low alpha value border (particle-def_larger_coverage.png). This means we can make the quad half as big in the vertex shader and achieve the same look.

Also instead of using the built-in Default-Particle, a new one has been added identical to it but only containing an alpha channel.